refactor: extract parseInitialViewState helper
