+++ humans-globe/components/footsteps/hooks/useGlobeViewState.ts
@@ -1,6 +1,9 @@
 'use client';
 import { useCallback, useEffect, useRef, useState } from 'react';
 import useDebouncedFlag from '@/components/footsteps/hooks/useDebouncedFlag';
+import parseInitialViewState, {
+  type ViewState,
+} from '@/lib/parseInitialViewState';
 
 type BasicViewState = {
   longitude: number;
@@ -10,49 +13,14 @@ type BasicViewState = {
   bearing?: number;
 };
 
-interface ViewState {
-  longitude: number;
-  latitude: number;
-  zoom: number;
-  pitch: number;
-  bearing: number;
-}
-
 export default function useGlobeViewState() {
   // Read optional initial state from URL query on first mount for QA/debug convenience
   // Example: ?lat=28.6&lon=77.2&zoom=8
-  const getInitialViewState = (): ViewState => {
-    try {
-      if (
-        typeof window !== 'undefined' &&
-        window.location &&
-        window.location.search
-      ) {
-        const sp = new URLSearchParams(window.location.search);
-        const lat = Number(sp.get('lat'));
-        const lon = Number(sp.get('lon') || sp.get('lng'));
-        const zoom = Number(sp.get('zoom') || sp.get('z'));
-        return {
-          longitude: Number.isFinite(lon) ? lon : 0,
-          latitude: Number.isFinite(lat) ? lat : 20,
-          zoom: Number.isFinite(zoom) ? zoom : 1.5,
-          pitch: 0,
-          bearing: 0,
-        };
-      }
-    } catch {
-      // ignore and fall back to defaults
-    }
-    return {
-      longitude: 0,
-      latitude: 20,
-      zoom: 1.5,
-      pitch: 0,
-      bearing: 0,
-    };
-  };
-
-  const [viewState, setViewState] = useState<ViewState>(getInitialViewState);
+  const [viewState, setViewState] = useState<ViewState>(() =>
+    parseInitialViewState(
+      typeof window !== 'undefined' ? window.location.search : '',
+    ),
+  );
   const viewStateRef = useRef<ViewState>(viewState);
 
   const [isZooming, triggerZoom] = useDebouncedFlag(800);
